% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{get_help_text}
\alias{get_help_text}
\title{Get help text from a string}
\usage{
get_help_text(help_string)
}
\arguments{
\item{help_string}{A string containing help text to be parsed.}
}
\value{
A string containing the help text.
}
\description{
This function takes a string of help text, parses it, and
returns the help text.
}
\examples{
res <- get_help_text("print")
cat(res)
}
