% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{inspect_object}
\alias{inspect_object}
\title{Inspect an object}
\usage{
inspect_object(x, name = deparse(substitute(x)))
}
\arguments{
\item{x}{The object to inspect.}

\item{name}{The name of the object.}
}
\description{
This function takes an object and returns a detailed summary of
its structure and properties.
}
\value{
Prints a comprehensive analysis of the provided object to the console, including 
its class, type, structure, dimensions, attributes, and additional details for 
special object types like S4 or R6. No return value, called for side effects.
}
\examples{
a = c(1,2,3)
res <- inspect_object(a)
cat(res)
}
