% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{run_r_string}
\alias{run_r_string}
\title{Run R code from a string}
\usage{
run_r_string(code_string)
}
\arguments{
\item{code_string}{A string containing R code to be executed.}
}
\value{
A list containing the output and the result of the evaluated code.
}
\description{
This function takes a string of R code, evaluates it, and
captures the output.
}
\examples{
res <- run_r_string("sum(1:10)")
cat("Input:", paste(res$output, collapse="\n"), "\n")
cat("Result:", res$result, "\n")
}
